fix: harden reload paths and prevent hangs, leaks, and use-after-free
Fix uint16 TAC range wrap on SIGHUP, S6a CLR/IDR error_out loops, SGWC
s5c modify_action UAF, strdup YAML-backed reload strings, ICS PDU leaks,
deferred config document retirement, and volatile reload counters.

+++ src/mme/mme-reload-lists.c
@@ -182,7 +182,7 @@ static int reload_served_tai_add_one(
 static int reload_served_tai_add_range(
         const ogs_plmn_id_t *plmn_id, uint16_t start, uint16_t end)
 {
-    uint16_t tac;
+    uint32_t tac;
     int added = 0;
 
     ogs_assert(plmn_id);
@@ -191,7 +191,7 @@ static int reload_served_tai_add_range(
         return 0;
 
     for (tac = start; tac <= end; tac++)
-        added += reload_served_tai_add_one(plmn_id, tac);
+        added += reload_served_tai_add_one(plmn_id, (uint16_t)tac);
 
     return added;
 }
